[#1|#8] - Closed issues #1 and #8 'Buy and sell coins endpoints'

+++ src/database/migrations/2021_05_11_104022_create_wallets_table.php
@@ -16,7 +16,7 @@ class CreateWalletsTable extends Migration
         Schema::create('wallets', function (Blueprint $table) {
             $table->string('wallet_id')->primary();
             $table->string('user_id');
-            $table->float('balance_usd');
+            $table->double('balance_usd', 32, 2);
 
             $table->foreign('user_id')->references('user_id')->on('users');
         });
